Rebuild for Python 3.7, GCC 7, R 3.5.1, openBLAS 0.3.2

+++ recipe/run_test.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+set -x
+
 echo "CONDA_PY:$CONDA_PY"
 export CONDA_PY=`python -c "import sys;print('%s%s'%sys.version_info[:2])"`
 echo "CONDA_PY:$CONDA_PY"
@@ -9,8 +11,10 @@ if [[ "$CONDA_PY" != "27" ]]; then
     grep "built-in method kill" gdb_output
 fi
 # Unfortunately python 3.6 package does not have enough debug info for py-bt
-if [[ "$CONDA_PY" != "36" ]]; then
+if [[ "$CONDA_PY" != "36" && "$CONDA_PY" != "27" ]]; then
     grep "line 3" gdb_output
     grep "process_to_debug.py" gdb_output
     grep 'os.kill(os.getpid(), signal.SIGSEGV)' gdb_output
 fi
+
+grep "Program received signal SIGSEGV" gdb_output
